Added PhysX. Updated premake.lua. PhysX implementation is in progress. Added Basic Project static class. Now C#::OnCreate is ran after all scripts were created. Added OnPhysicsUpdate to C#. Added FileSystem::Read & FileSystem::Write. Added DataBuffer.

+++ Eagle/src/Eagle/Utils/Utils.cpp
@@ -24,11 +24,11 @@ namespace Eagle::Utils
 	std::string toUtf8(const std::wstring& str)
 	{
 		std::string ret;
-		int len = WideCharToMultiByte(CP_UTF8, 0, str.c_str(), str.length(), NULL, 0, NULL, NULL);
+		int len = WideCharToMultiByte(CP_UTF8, 0, str.c_str(), (int)str.length(), NULL, 0, NULL, NULL);
 		if (len > 0)
 		{
 			ret.resize(len);
-			WideCharToMultiByte(CP_UTF8, 0, str.c_str(), str.length(), &ret[0], len, NULL, NULL);
+			WideCharToMultiByte(CP_UTF8, 0, str.c_str(), (int)str.length(), &ret[0], len, NULL, NULL);
 		}
 		return ret;
 	}
